Refactor: telemetry router (#1061)
