+++ backend/src/providers/telemetry.ts
@@ -1,61 +1,30 @@
-import geoip from "geoip-lite";
-import { z } from "zod";
-import {
-  maybeCreateTelemetryTable,
-  maybeSendTelemetry,
-} from "../database/queries";
-import { validators } from "../router/validators";
+import { telemetryWriteDatabase } from "../database/databases";
 
 // Skip initializing Telemetry database if the backend is not configured to
 // save telemety data (it is absolutely fine for local development)
 export const setup = async () => {
-  await maybeCreateTelemetryTable();
-};
-
-export const sendTelemetry = async (
-  nodeInfo: z.infer<typeof validators.telemetryRequest>
-): Promise<void> => {
-  if (!nodeInfo.hasOwnProperty("agent")) {
-    // This seems to be an old format, and all our nodes should support the new
-    // Telemetry format as of 2020-04-14, so we just ignore those old Telemetry
-    // reports.
+  if (!telemetryWriteDatabase) {
     return;
   }
-
-  // TODO update validators list ones per epoch
-
-  // const stakingNodesList = await getStakingNodesList();
-  // const stakingNode = stakingNodesList.get(nodeInfo.chain.account_id);
-  // we want to validate "active" validators only
-  // const isValidator = stakingNode?.stakingStatus === "active";
-
-  // TODO: fix the signature verification. Given that the JSON payload is signed, it creates a challenge to serialize the same JSON string without `signature` field.
-  /*
-  if (isValidator) {
-    const messageJSON = {
-      agent: nodeInfo.agent,
-      system: nodeInfo.system,
-      chain: nodeInfo.chain,
-    };
-    const message = new TextEncoder().encode(messageJSON);
-    const publicKey = utils.PublicKey.from(stakingNode.public_key);
-    const isVerified = nacl.sign.detached.verify(
-      message,
-      utils.serialize.base_decode(
-        telemetryInfo.signature.substring(8, telemetryInfo.signature.length)
-      ),
-      publicKey.data
-    );
-
-    if (!isVerified) {
-      console.warn(
-        "We ignore fake telemetry data about validation node: ",
-        nodeInfo
-      );
-      return;
-    }
-  }
-  */
-  const geo = geoip.lookup(nodeInfo.ip_address);
-  return maybeSendTelemetry(nodeInfo, geo);
+  await telemetryWriteDatabase.schema
+    .createTable("nodes")
+    .ifNotExists()
+    .addColumn("ip_address", "varchar(255)", (col) => col.notNull())
+    .addColumn("moniker", "varchar(255)", (col) => col.notNull())
+    .addColumn("account_id", "varchar(255)", (col) => col.notNull())
+    .addColumn("node_id", "varchar(255)", (col) => col.notNull().primaryKey())
+    .addColumn("last_seen", "timestamptz", (col) => col.notNull())
+    .addColumn("last_height", "bigint", (col) => col.notNull())
+    .addColumn("agent_name", "varchar(255)", (col) => col.notNull())
+    .addColumn("agent_version", "varchar(255)", (col) => col.notNull())
+    .addColumn("agent_build", "varchar(255)", (col) => col.notNull())
+    .addColumn("peer_count", "varchar(255)", (col) => col.notNull())
+    .addColumn("is_validator", "boolean", (col) => col.notNull())
+    .addColumn("last_hash", "varchar(255)", (col) => col.notNull())
+    .addColumn("signature", "varchar(255)", (col) => col.notNull())
+    .addColumn("status", "varchar(255)", (col) => col.notNull())
+    .addColumn("latitude", "numeric(8, 6)")
+    .addColumn("longitude", "numeric(9, 6)")
+    .addColumn("city", "varchar(255)")
+    .execute();
 };
